backend/notification_service/worker/requirements.txt
aio_pika==8.3.0
pydantic==1.10.4
grpcio-tools==1.51.1
orjson==3.8.5
jinja2==3.1.2
backoff==2.2.1
requests==2.28.2
flake8==6.0.0
isort==5.12.0